// apps/io-func-sign-user/src/infra/http/models/SignatureRequestDetailView.ts
/**
 * Do not edit this file it is auto-generated by io-utils / gen-api-models.
 * See https://github.com/pagopa/io-utils
 */
/* eslint-disable  */

import { Id } from "./Id";
import { SignatureRequestStatus } from "./SignatureRequestStatus";
import { DossierTitle } from "./DossierTitle";
import { Timestamp } from "./Timestamp";
import { DocumentDetailView } from "./DocumentDetailView";
import * as t from "io-ts";
import { IssuerEnvironment } from "./IssuerEnvironment";
import { EmailString } from "@pagopa/ts-commons/lib/strings";
import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";

// required attributes
const SignatureRequestDetailViewIssuerR = t.interface({
  email: EmailString,

  description: NonEmptyString,

  environment: IssuerEnvironment
});

// optional attributes
const SignatureRequestDetailViewIssuerO = t.partial({});

export const SignatureRequestDetailViewIssuer = t.exact(
  t.intersection(
    [SignatureRequestDetailViewIssuerR, SignatureRequestDetailViewIssuerO],
    "SignatureRequestDetailViewIssuer"
  )
);

export type SignatureRequestDetailViewIssuer = t.TypeOf<
  typeof SignatureRequestDetailViewIssuer
>;

// required attributes
const SignatureRequestDetailViewR = t.interface({
  id: Id,

  status: SignatureRequestStatus,

  issuer: SignatureRequestDetailViewIssuer,

  dossier_id: Id,

  dossier_title: DossierTitle,

  signer_id: Id,

  expires_at: Timestamp,

  documents: t.readonlyArray(DocumentDetailView, "array of DocumentDetailView"),

  created_at: Timestamp,

  updated_at: Timestamp
});

// optional attributes
const SignatureRequestDetailViewO = t.partial({
  qr_code_url: t.string
});

export const SignatureRequestDetailView = t.exact(
  t.intersection(
    [SignatureRequestDetailViewR, SignatureRequestDetailViewO],
    "SignatureRequestDetailView"
  )
);

export type SignatureRequestDetailView = t.TypeOf<
  typeof SignatureRequestDetailView
>;